requirements.txt
pip
lxml
tqdm
requests
colorama
PyYAML
numba
numpy
llvmlite
cffi
scipy
setuptools
Cython
pooch
packaging
appdirs
matplotlib
joblib
certifi
chardet
SoundFile
audioread
resampy
decorator
scikit-learn
praatio
six
threadpoolctl
Pillow
urllib3
idna
pyparsing
pycparser
pysrt
pyttsx3
audiotsm2
jieba
pypinyin
pyworld
pydub
pypiwin32
pypinyin-dict